Reduce CPU usage during transmission
- Skip video frame decoding between thumbnails (only decode every 120th frame)
- Add condition variables to dataqueue replacing busy-wait polling loops
- Increase error sleep delays from 100us to 10ms in SRT receive/send paths

+++ source/srthub.c
@@ -240,44 +240,33 @@ static int receive_frame(uint8_t *sample, int sample_size, int sample_type, uint
     }
 
     if (sample_type == STREAM_TYPE_H264 || sample_type == STREAM_TYPE_HEVC || sample_type == STREAM_TYPE_MPEG2) {
-        //if (sample_flags == 1)
-        {
-            /*fprintf(stderr,"received frame (%d/%d): source=%d, sub_source=%d, type=0x%x, corruption_count=%ld, size=%d\n",
-              source+1, muxstreams, source, sub_source, sample_type, corruption_count, sample_size);*/
+        srtcore->thumbnail_frame_counter++;
+        
+        #define THUMBNAIL_FRAME_INTERVAL 120
+        if ((srtcore->thumbnail_frame_counter % THUMBNAIL_FRAME_INTERVAL) != 0) {
+            return 0;
+        }
 
-            /*if (dataqueue_get_size(srtcore->thumbnailqueue) > 0) {
-                fprintf(stderr,"received frame: thumbnail queue already has a sample... waiting\n");
-                return 0;
-            }*/
-
-            int i;
-            for (i = 0; i < sample_size-3; i++) {
-                if (sample[i] == 0x00 && sample[i+1] == 0x00 && sample[i+2] == 0x01) {
-                    //fprintf(stderr,"received frame: start code found pos %4d: type=0x%x\n", i, sample[i+3]);
-                }
-            }
-
-            msg = (dataqueue_message_struct*)memory_take(srtcore->msgpool, threadid);
-            if (msg) {
-                uint8_t *buffer = (uint8_t*)memory_take(srtcore->videopool, threadid);
-                if (buffer) {
-                    memcpy(buffer, sample, sample_size);
-                    msg->buffer = (void*)buffer;
-                    msg->buffer_size = sample_size;
-                    msg->buffer_type = sample_type;
-                    msg->flags = muxstreams;
-                    msg->stream_index = source;
-                    msg->source_discontinuity = corruption_count;
-                    dataqueue_put_front(srtcore->thumbnailqueue, msg);
-                    msg = NULL;
-                } else {
-                    fprintf(stderr,"received frame: thumbnail buffers exhausted\n");
-                    memory_return(srtcore->msgpool, msg);
-                    msg = NULL;
-                }
+        msg = (dataqueue_message_struct*)memory_take(srtcore->msgpool, threadid);
+        if (msg) {
+            uint8_t *buffer = (uint8_t*)memory_take(srtcore->videopool, threadid);
+            if (buffer) {
+                memcpy(buffer, sample, sample_size);
+                msg->buffer = (void*)buffer;
+                msg->buffer_size = sample_size;
+                msg->buffer_type = sample_type;
+                msg->flags = muxstreams;
+                msg->stream_index = source;
+                msg->source_discontinuity = corruption_count;
+                dataqueue_put_front(srtcore->thumbnailqueue, msg);
+                msg = NULL;
             } else {
-                fprintf(stderr,"received frame: msg buffers exhausted, thumbnailqueue=%d\n", dataqueue_get_size(srtcore->msgpool));
+                fprintf(stderr,"received frame: thumbnail buffers exhausted\n");
+                memory_return(srtcore->msgpool, msg);
+                msg = NULL;
             }
+        } else {
+            fprintf(stderr,"received frame: msg buffers exhausted, thumbnailqueue=%d\n", dataqueue_get_size(srtcore->msgpool));
         }
     }
 
@@ -495,9 +484,9 @@ static void *srt_receiver_thread_listener(void *context)
                     send_restart_message(srtcore);
                     goto cleanup_srt_receiver_thread_listener;
                 }
-                usleep(100);
+                usleep(10000);
             } else if (recvbytes == 0) {
-                usleep(100);
+                usleep(10000);
             } else {
                 int cp;
                 int tp;
@@ -809,9 +798,9 @@ static void *srt_receiver_thread_caller(void *context)
             } else {
                 fprintf(stderr,"srt_receiver_thread_caller: SRT unknown error: %s\n", srt_getlasterror_str());
             }
-            usleep(100);
+            usleep(10000);
         } else if (recvbytes == 0) {
-            usleep(100);
+            usleep(10000);
         } else {
             int cp;
             int tp;
@@ -961,16 +950,7 @@ static void *srt_server_worker_output_thread(void *context)
 
     clock_gettime(CLOCK_MONOTONIC, &server_time_start);
     while (srtcore->srt_server_worker_thread_running[thread]) {
-        pthread_mutex_lock(srtcore->srtserverlock);
-        msg = (dataqueue_message_struct*)dataqueue_take_back(srtcore->srtserverqueue[thread]);
-        pthread_mutex_unlock(srtcore->srtserverlock);
-
-        while (!msg && srtcore->srt_server_worker_thread_running[thread]) {
-            usleep(1000);
-            pthread_mutex_lock(srtcore->srtserverlock);
-            msg = (dataqueue_message_struct*)dataqueue_take_back(srtcore->srtserverqueue[thread]);
-            pthread_mutex_unlock(srtcore->srtserverlock);
-        }
+        msg = (dataqueue_message_struct*)dataqueue_take_back_wait(srtcore->srtserverqueue[thread], &srtcore->srt_server_worker_thread_running[thread]);
 
         if (!srtcore->srt_server_worker_thread_running[thread]) {
             if (msg) {
@@ -1031,9 +1011,9 @@ static void *srt_server_worker_output_thread(void *context)
             } else {
                 fprintf(stderr,"srt_server_worker_output_thread: SRT unknown error: %s\n", srt_getlasterror_str());
             }
-            usleep(100);
+            usleep(10000);
         } else if (sent_bytes == 0) {
-            usleep(100);
+            usleep(10000);
         } else {
             total_packets_sent++;
             total_bytes_sent += sent_bytes;
@@ -1447,9 +1427,9 @@ retry_srt_server_push_connection:
                 fprintf(stderr,"srt_server_thread_push: SRT unknown error: %s\n", srt_getlasterror_str());
                 // need to handle this error
             }
-            usleep(100);
+            usleep(10000);
         } else if (sent_bytes == 0) {
-            usleep(100);
+            usleep(10000);
         } else {
             total_packets_sent++;
             total_bytes_sent += sent_bytes;
@@ -1827,42 +1807,37 @@ static void *udp_server_thread(void *context)
     sprintf(statsfilename,"/opt/srthub/status/udp_server_%d.json", srtcore->session_identifier);
 
     clock_gettime(CLOCK_MONOTONIC, &stats_start);
+    clock_gettime(CLOCK_MONOTONIC, &signal_check_start);
     while (srtcore->udp_server_thread_running) {
-        msg = (dataqueue_message_struct*)dataqueue_take_back(srtcore->udpserverqueue);
+        msg = (dataqueue_message_struct*)dataqueue_take_back_wait(srtcore->udpserverqueue, &srtcore->udp_server_thread_running);
 
-        if (!msg) {
-            clock_gettime(CLOCK_MONOTONIC, &signal_check_start);
-        }
-        while (!msg && srtcore->udp_server_thread_running) {
-            usleep(1000);
-            msg = (dataqueue_message_struct*)dataqueue_take_back(srtcore->udpserverqueue);
-            if (!msg) {
-                clock_gettime(CLOCK_MONOTONIC, &signal_check_stop);
-                diff = realtime_clock_difference(&signal_check_stop, &signal_check_start) / 1000;
-                if (diff >= 2000) {  // 2 second timeout
-                    FILE *statsfile = fopen(statsfilename,"wb");
-                    if (statsfile) {
-                        fprintf(statsfile,"{\n");
-                        fprintf(statsfile,"    \"udp-output-address\":\"%s\",\n", udpdata->destination_address);
-                        fprintf(statsfile,"    \"udp-output-port\":%d,\n", udpdata->destination_port);
-                        fprintf(statsfile,"    \"udp-output-interface\":\"%s\",\n", udpdata->interface_name);
-                        fprintf(statsfile,"    \"udp-output-ttl\":%d,\n", udpdata->ttl);
-                        fprintf(statsfile,"    \"udp-output-active\":0,\n");
-                        fprintf(statsfile,"    \"total-bytes-sent\":%ld,\n", total_bytes_sent);
-                        fprintf(statsfile,"    \"total-packets-sent\":%ld,\n", total_packets_sent);
-                        fprintf(statsfile,"    \"last-buffer-size\":%d,\n", 0);
-                        fprintf(statsfile,"    \"multicast-output\":%d,\n", multicast_output);
-                        fprintf(statsfile,"    \"udpserver-queue\":%d\n", dataqueue_get_size(srtcore->udpserverqueue));
-                        fprintf(statsfile,"}\n");
-                        fclose(statsfile);
-                    }
-                    if (!signal_outage_flagged) {
-                        signal_outage_flagged = 1;
-
-                        send_signal(srtcore, SIGNAL_NO_DATA, "No Data on SRT Connection");
-                    }
+        if (!msg && srtcore->udp_server_thread_running) {
+            clock_gettime(CLOCK_MONOTONIC, &signal_check_stop);
+            diff = realtime_clock_difference(&signal_check_stop, &signal_check_start) / 1000;
+            if (diff >= 2000) {
+                FILE *statsfile = fopen(statsfilename,"wb");
+                if (statsfile) {
+                    fprintf(statsfile,"{\n");
+                    fprintf(statsfile,"    \"udp-output-address\":\"%s\",\n", udpdata->destination_address);
+                    fprintf(statsfile,"    \"udp-output-port\":%d,\n", udpdata->destination_port);
+                    fprintf(statsfile,"    \"udp-output-interface\":\"%s\",\n", udpdata->interface_name);
+                    fprintf(statsfile,"    \"udp-output-ttl\":%d,\n", udpdata->ttl);
+                    fprintf(statsfile,"    \"udp-output-active\":0,\n");
+                    fprintf(statsfile,"    \"total-bytes-sent\":%ld,\n", total_bytes_sent);
+                    fprintf(statsfile,"    \"total-packets-sent\":%ld,\n", total_packets_sent);
+                    fprintf(statsfile,"    \"last-buffer-size\":%d,\n", 0);
+                    fprintf(statsfile,"    \"multicast-output\":%d,\n", multicast_output);
+                    fprintf(statsfile,"    \"udpserver-queue\":%d\n", dataqueue_get_size(srtcore->udpserverqueue));
+                    fprintf(statsfile,"}\n");
+                    fclose(statsfile);
                 }
+                if (!signal_outage_flagged) {
+                    signal_outage_flagged = 1;
+                    send_signal(srtcore, SIGNAL_NO_DATA, "No Data on SRT Connection");
+                }
+                clock_gettime(CLOCK_MONOTONIC, &signal_check_start);
             }
+            continue;
         }
 
         if (!srtcore->udp_server_thread_running) {
@@ -1985,13 +1960,7 @@ static void *srthub_audio_thread(void *context)
 
     clock_gettime(CLOCK_MONOTONIC, &audio_start);
     while (srtcore->audio_decode_thread_running[audio_stream]) {
-        //fprintf(stderr,"srthub_audio_thread: srtcore=%p\n", srtcore);
-        //fprintf(stderr,"srthub_audio_thread: queue=%p audio_stream=%d\n", srtcore->audiodecodequeue[audio_stream], audio_stream);
-        msg = (dataqueue_message_struct*)dataqueue_take_back(srtcore->audiodecodequeue[audio_stream]);
-        while (!msg && srtcore->audio_decode_thread_running[audio_stream]) {
-            usleep(1000);
-            msg = (dataqueue_message_struct*)dataqueue_take_back(srtcore->audiodecodequeue[audio_stream]);
-        }
+        msg = (dataqueue_message_struct*)dataqueue_take_back_wait(srtcore->audiodecodequeue[audio_stream], &srtcore->audio_decode_thread_running[audio_stream]);
 
         if (!srtcore->audio_decode_thread_running[audio_stream]) {
             if (msg) {
@@ -2189,11 +2158,7 @@ static void *srthub_thumbnail_thread(void *context)
 
     memset(corruptiontimedate, 0, sizeof(corruptiontimedate));
     while (srtcore->thumbnail_thread_running) {
-        msg = (dataqueue_message_struct*)dataqueue_take_back(srtcore->thumbnailqueue);
-        while (!msg && srtcore->thumbnail_thread_running) {
-            usleep(1000);
-            msg = (dataqueue_message_struct*)dataqueue_take_back(srtcore->thumbnailqueue);
-        }
+        msg = (dataqueue_message_struct*)dataqueue_take_back_wait(srtcore->thumbnailqueue, &srtcore->thumbnail_thread_running);
 
         if (!srtcore->thumbnail_thread_running) {
             if (msg) {
@@ -2327,11 +2292,7 @@ static void *srthub_thumbnail_thread(void *context)
                     av_image_alloc(output_data, output_stride, THUMBNAIL_WIDTH, THUMBNAIL_HEIGHT, output_format, 1);
                 }
 
-                if ((decoded_frame_count % 120)==0) {
-                    /*
-                    fprintf(stderr,"srt_thumbnail_thread: decoded video frame, resolution is %d x %d\n",
-                            frame_width, frame_height);
-                    */
+                {
                     char codec[MAX_STRING_SIZE];
                     if (buffer_type == STREAM_TYPE_H264) {
                         sprintf(codec,"h264");
